TimecodePanel リセットボタンのテスト追加
理由:
- リセットボタンが Date.now() を startDateTime に設定することのテストがなかった

変更内容:
- timecodeOverlay.test.ts → .tsx にリネーム（JSX コンポーネントテスト追加のため）
- TimecodePanel のリセットボタンが onChange を Date.now() の startDateTime で呼ぶことを検証
